Update validators.R
I added a validator testing if datetime_ is in the format yyyy-mm-dd hh:mm (e.g. one has to add the time in the column with prefilled date) and a validator testing for the right order of events (e.g. caught must be before bleeding). Both work, but the later gives a warning: "POSIXlt column type detected and converted to POSIXct. We do not recommend use of POSIXlt at all because it uses 40 bytes to store one date." of which I could not get rid of.